move authenication-related methods their own helper
