+++ app/controllers/application_controller.rb
@@ -2,6 +2,7 @@ class ApplicationController < ActionController::Base
   include Pundit    # For authorization checks
   include ReturnToHelper
   include MarkdownHelper
+  include AuthenticationHelper
 
   helper ValueHelper
   add_template_helper ClientHelper
@@ -36,7 +37,7 @@ class ApplicationController < ActionController::Base
       record
     end
   end
-  
+
   # Override Pundit to account for proposal gymnastics
   def policy(record)
     super(self.authorizing_object(record))
@@ -51,25 +52,6 @@ class ApplicationController < ActionController::Base
   end
 
   private
-
-  def current_user
-    @current_user ||= User.find_or_create_by(email_address: session[:user]['email']) if session[:user] && session[:user]['email']
-  end
-
-  def sign_in(user)
-    session[:user] ||= {}
-    session[:user]['email'] = user.email_address
-    @current_user = user
-  end
-
-  def sign_out
-    reset_session
-    @current_user = nil
-  end
-
-  def signed_in?
-    !!current_user
-  end
 
   def authenticate_user!
     unless signed_in?
