+++ spec/rails_helper.rb
@@ -66,6 +66,7 @@ RSpec.configure do |config|
   config.include FeatureSpecHelper, type: :feature
 
   config.include EnvironmentSpecHelper, type: :controller
+  config.include AuthenticationHelper, type: :helper
 
   # Much of the config here pieced together from
   # http://stackoverflow.com/questions/8178120/capybara-with-js-true-causes-test-to-fail/28083267
